added ZeroizableSecretKey and moved tests
