+++ crates/bridge-exec/src/handler.rs
@@ -355,59 +355,3 @@ where
         self.keypair.zeroize();
     }
 }
-
-#[cfg(test)]
-mod tests {
-
-    use std::sync::{
-        atomic::{AtomicBool, Ordering},
-        Arc,
-    };
-
-    use bitcoin::{bip32::Xpriv, secp256k1::SECP256K1};
-    use strata_primitives::keys::ZeroizableKeypair;
-    use zeroize::Zeroize;
-
-    // What's better than bacon? bacon^24 of course
-    // Thix xpriv was generated by the bacon^24 mnemonic
-    // Don't use this in production!
-    const XPRIV_STR: &str = "tprv8ZgxMBicQKsPeh9dSitM82FU7Fz3ZgPkKmmovAr2aqwauAMVgjcEkZBb2etBtRPZ8XYVm7shxcKwVaDus7T5kauJXVsqAfzM4Tty13rRjAG";
-
-    #[test]
-    fn test_erase_keypair() {
-        // Create a wrapper struct including keypair
-        struct TestWrapper {
-            inner: ZeroizableKeypair,
-            flag: Arc<AtomicBool>,
-        }
-
-        // Create an atomic flag to track if zeroize was called
-        let was_zeroized = Arc::new(AtomicBool::new(false));
-        let was_zeroized_clone = Arc::clone(&was_zeroized);
-
-        impl Drop for TestWrapper {
-            fn drop(&mut self) {
-                // Get the current value before the inner value is dropped
-                let bytes = self.inner.secret_bytes();
-                self.inner.zeroize();
-                // The inner Keypair will be dropped after this,
-                // triggering zeroization
-                // NOTE: SecretKey::non_secure_erase writes `1`s to the memory.
-                self.flag.store(bytes != [1u8; 32], Ordering::Relaxed);
-            }
-        }
-
-        {
-            let xpriv = XPRIV_STR.parse::<Xpriv>().unwrap();
-            let keypair = xpriv.to_keypair(SECP256K1);
-
-            let _ = TestWrapper {
-                inner: keypair.into(),
-                flag: was_zeroized_clone,
-            };
-        }
-
-        // Check if zeroization occurred
-        assert!(was_zeroized.load(Ordering::Relaxed));
-    }
-}
