+++ crates/sequencer/Cargo.toml
@@ -25,7 +25,9 @@ tracing.workspace = true
 zeroize.workspace = true
 
 [dev-dependencies]
+bitcoin.workspace = true
 strata-test-utils.workspace = true
+strata-key-derivation.workspace = true
 
 [lints]
 rust.missing_docs = "warn"
